@@ -23,6 +23,7 @@ import org.springframework.boot.info.BuildProperties;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import java.util.concurrent.Executors;
@@ -30,17 +31,18 @@ import java.util.concurrent.TimeUnit;
 
 @Controller
 @RequiredArgsConstructor
+@RequestMapping("/")
 public class ApplicationRestController {
 
     private final BuildProperties buildProperties;
 
-    @GetMapping("/")
+    @GetMapping
     public String get(Model model) {
         model.addAttribute("buildProperties", buildProperties);
         return "index";
     }
 
-    @GetMapping("/shutdown")
+    @GetMapping("shutdown")
     @ResponseBody
     public String shutdown() {
         Executors.newSingleThreadScheduledExecutor().schedule(() -> System.exit(0), 3, TimeUnit.SECONDS);
