@@ -19,12 +19,10 @@
 package ru.investbook.parser.psb;
 
 import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import org.apache.poi.hssf.usermodel.HSSFWorkbook;
 import org.apache.poi.ss.usermodel.Workbook;
-import org.apache.poi.xssf.usermodel.XSSFWorkbook;
-import ru.investbook.parser.BrokerReport;
+import ru.investbook.parser.AbstractBrokerReport;
 import ru.investbook.parser.table.ReportPage;
+import ru.investbook.parser.table.TableCellAddress;
 import ru.investbook.parser.table.excel.ExcelSheet;
 
 import java.io.IOException;
@@ -33,31 +31,15 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.time.Instant;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.ZoneId;
-import java.time.format.DateTimeFormatter;
 import java.time.temporal.ChronoUnit;
 
-@EqualsAndHashCode(of = "path")
-public class PsbBrokerReport implements BrokerReport {
-    private static final DateTimeFormatter dateFormatter = DateTimeFormatter.ofPattern("dd.MM.yyyy");
-    private static final DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("dd.MM.yyyy HH:mm:ss");
+@EqualsAndHashCode(callSuper = true)
+public class PsbBrokerReport extends AbstractBrokerReport {
+    public static final String UNIQ_TEXT = "Брокер: ПАО \"Промсвязьбанк\"";
     private static final String PORTFOLIO_MARKER = "Договор №:";
     private static final String REPORT_DATE_MARKER = "ОТЧЕТ БРОКЕРА";
-    private static final int LAST_TRADE_HOUR = 19;
 
     private final Workbook book;
-    @Getter
-    private final ReportPage reportPage;
-    @Getter
-    private final String portfolio;
-    @Getter
-    private final Path path;
-    @Getter
-    private final Instant reportEndDateTime;
-    @Getter
-    private final ZoneId reportZoneId = ZoneId.of("Europe/Moscow");
 
     public PsbBrokerReport(String excelFileName) throws IOException {
         this(Paths.get(excelFileName));
@@ -67,55 +49,40 @@ public class PsbBrokerReport implements BrokerReport {
         this(report.getFileName().toString(), Files.newInputStream(report));
     }
 
-    public PsbBrokerReport(String excelFileName, InputStream is) throws IOException {
-        this.path = Paths.get(excelFileName);
+    public PsbBrokerReport(String excelFileName, InputStream is) {
         this.book = getWorkBook(excelFileName, is);
-        this.reportPage = new ExcelSheet(book.getSheetAt(0));
-        this.portfolio = getPortfolio(this.reportPage);
-        this.reportEndDateTime = getReportEndDateTime(this.reportPage);
+        ReportPage reportPage = new ExcelSheet(book.getSheetAt(0));
+        checkReportFormat(excelFileName, reportPage);
+        setPath(Paths.get(excelFileName));
+        setReportPage(reportPage);
+        setPortfolio(getPortfolio(reportPage));
+        setReportEndDateTime(getReportEndDateTime(reportPage));
     }
 
-    private Workbook getWorkBook(String excelFileName, InputStream is) throws IOException {
-        if (excelFileName.endsWith(".xls")) {
-            return new HSSFWorkbook(is); // constructor close zis
-        } else {
-            return new XSSFWorkbook(is);
+    public static void checkReportFormat(String excelFileName, ReportPage reportPage) {
+        if (reportPage.find(UNIQ_TEXT, 3, 4) == TableCellAddress.NOT_FOUND) {
+            throw new RuntimeException("В файле " + excelFileName + " не содержится отчет брокера Промсвязьбанк");
         }
     }
 
     private static String getPortfolio(ReportPage reportPage) {
         try {
             String value = String.valueOf(reportPage.getNextColumnValue(PORTFOLIO_MARKER));
-            if (value != null) {
-                return value.contains("/") ? value.split("/")[0] : value;
-            }
+            return value.contains("/") ? value.split("/")[0] : value;
+        } catch (Exception e) {
             throw new IllegalArgumentException(
                     "В отчете не найден номер договора по заданному шаблону '" + PORTFOLIO_MARKER + " XXX'");
-        } catch (Exception e) {
-            throw new RuntimeException("Ошибка поиска номера Брокерского счета в отчете");
         }
     }
 
     private Instant getReportEndDateTime(ReportPage reportPage) {
         try {
             String value = String.valueOf(reportPage.getNextColumnValue(REPORT_DATE_MARKER));
-            if (value != null) {
-                return convertToInstant(value.split(" ")[3])
-                        .plus(LAST_TRADE_HOUR, ChronoUnit.HOURS);
-            }
+            return convertToInstant(value.split(" ")[3])
+                    .plus(LAST_TRADE_HOUR, ChronoUnit.HOURS);
+        } catch (Exception e) {
             throw new IllegalArgumentException(
                     "Не найдена дата отчета по заданному шаблону '" + REPORT_DATE_MARKER + " XXX'");
-        } catch (Exception e) {
-            throw new RuntimeException("Ошибка поиска даты отчета");
-        }
-    }
-
-    public Instant convertToInstant(String value) {
-        value = value.trim();
-        if (value.contains(":")) {
-            return LocalDateTime.parse(value, PsbBrokerReport.dateTimeFormatter).atZone(getReportZoneId()).toInstant();
-        } else {
-            return LocalDate.parse(value, PsbBrokerReport.dateFormatter).atStartOfDay(getReportZoneId()).toInstant();
         }
     }
 
