@@ -29,8 +29,8 @@ public enum CashFlowType {
     COMMISSION(3), // комиссия
     AMORTIZATION(4),    // амортизация
     REDEMPTION(5), // погашение номинала облигации
-    COUPON(6), // выплата купона
-    DIVIDEND(7), // выплата дивиденда
+    COUPON(6), // выплата купона (сумма до удержания налога)
+    DIVIDEND(7), // выплата дивиденда (сумма до удержания налога)
     DERIVATIVE_PROFIT(8),// вариационная маржа
     MARGIN(9), // гарантийное обеспечение
     TAX(10), // налог уплаченный
