@@ -42,6 +42,7 @@ import static ru.investbook.view.excel.TaxExcelTableHeader.*;
 public class TaxExcelTableFactory implements TableFactory {
     private final EventCashFlowRepository eventCashFlowRepository;
     private final EventCashFlowConverter eventCashFlowConverter;
+    private final ForeignExchangeRateTableFactory foreignExchangeRateTableFactory;
 
     @Override
     public Table create(Portfolio portfolio) {
@@ -56,16 +57,35 @@ public class TaxExcelTableFactory implements TableFactory {
                 .map(eventCashFlowConverter::fromEntity)
                 .collect(Collectors.toCollection(ArrayList::new));
 
-        for (EventCashFlow cash : cashFlows) {
-            Table.Record record = new Table.Record();
-            record.put(DATE, cash.getTimestamp());
-            record.put(TAX, Optional.ofNullable(cash.getValue())
-                    .map(BigDecimal::negate)
-                    .orElse(BigDecimal.ZERO));
-            record.put(CURRENCY, cash.getCurrency());
-            record.put(DESCRIPTION, cash.getDescription());
-            table.add(record);
+        cashFlows.stream()
+                .filter(cash -> !isDividendOrCouponTax(cash.getDescription()))
+                .forEach(cash -> addRecordToTable(table, cash));
+
+        if (!cashFlows.isEmpty()) {
+            foreignExchangeRateTableFactory.appendExchangeRates(table, CURRENCY_NAME, EXCHANGE_RATE);
         }
+
         return table;
+    }
+
+    private void addRecordToTable(Table table, EventCashFlow cash) {
+        Table.Record record = new Table.Record();
+        record.put(DATE, cash.getTimestamp());
+        record.put(TAX, Optional.ofNullable(cash.getValue())
+                .map(BigDecimal::negate)
+                .orElse(BigDecimal.ZERO));
+        record.put(CURRENCY, cash.getCurrency());
+        record.put(TAX_RUB, foreignExchangeRateTableFactory.cashConvertToRubExcelFormula(cash.getCurrency(),
+                TAX, EXCHANGE_RATE));
+        record.put(DESCRIPTION, cash.getDescription());
+        table.add(record);
+    }
+
+    static boolean isDividendOrCouponTax(String description) {
+        if (description == null) {
+            return false;
+        }
+        String lowercasedDescription = description.toLowerCase();
+        return lowercasedDescription.contains("дивиденд") || lowercasedDescription.contains("купон");
     }
 }
