@@ -38,7 +38,7 @@ public class PortfolioSecuritiesTableTest {
 
     @Test(dataProvider = "isin")
     void testIsin(String report, String firstIsin, String lastIsin) throws IOException {
-        List<Security> data = new PortfolioSecuritiesTable(new PsbBrokerReport(report)).getData();
+        List<Security> data = new SecuritiesTable(new PsbBrokerReport(report)).getData();
         assertEquals(data.get(0).getIsin(), firstIsin);
         assertEquals(data.get(data.size() - 1).getIsin(), lastIsin);
     }
